rename data exceptions and add a few data sanity checks
